requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.8,<5.0
django-environ>=0.8.1,<1.0
